整理http文件服务器相关代码
