+++ src/Http/HttpSession.h
@@ -27,45 +27,18 @@
 #define SRC_HTTP_HTTPSESSION_H_
 
 #include <functional>
-#include "Common/config.h"
-#include "Common/Parser.h"
 #include "Network/TcpSession.h"
-#include "Network/TcpServer.h"
 #include "Rtmp/RtmpMediaSource.h"
 #include "Rtmp/FlvMuxer.h"
 #include "HttpRequestSplitter.h"
 #include "WebSocketSplitter.h"
 #include "HttpCookieManager.h"
-#include "HttpBody.h"
-#include "Util/function_traits.h"
+#include "HttpFileManager.h"
 
 using namespace std;
 using namespace toolkit;
 
 namespace mediakit {
-
-/**
- * 该类实现与老代码的兼容适配
- */
-class HttpResponseInvokerImp{
-public:
-	typedef std::function<void(const string &codeOut, const StrCaseMap &headerOut, const HttpBody::Ptr &body)> HttpResponseInvokerLambda0;
-	typedef std::function<void(const string &codeOut, const StrCaseMap &headerOut, const string &body)> HttpResponseInvokerLambda1;
-
-	HttpResponseInvokerImp(){}
-	~HttpResponseInvokerImp(){}
-	template<typename C>
-	HttpResponseInvokerImp(const C &c):HttpResponseInvokerImp(typename function_traits<C>::stl_function_type(c)) {}
-	HttpResponseInvokerImp(const HttpResponseInvokerLambda0 &lambda);
-	HttpResponseInvokerImp(const HttpResponseInvokerLambda1 &lambda);
-
-	void operator()(const string &codeOut, const StrCaseMap &headerOut, const HttpBody::Ptr &body) const;
-	void operator()(const string &codeOut, const StrCaseMap &headerOut, const string &body) const;
-	void responseFile(const StrCaseMap &requestHeader,const StrCaseMap &responseHeader,const string &filePath) const;
-	operator bool();
-private:
-	HttpResponseInvokerLambda0 _lambad;
-};
 
 class HttpSession: public TcpSession,
                    public FlvMuxer,
@@ -88,9 +61,7 @@ public:
 	virtual void onRecv(const Buffer::Ptr &) override;
 	virtual void onError(const SockException &err) override;
 	virtual void onManager() override;
-
 	static string urlDecode(const string &str);
-	static const char* get_mime_type(const char* name);
 protected:
 	//FlvMuxer override
 	void onWrite(const Buffer::Ptr &data) override ;
@@ -145,26 +116,6 @@ private:
 	void sendResponse(const char *pcStatus, bool bClose, const char *pcContentType = nullptr,
 					  const HttpSession::KeyValue &header = HttpSession::KeyValue(),
                       const HttpBody::Ptr &body = nullptr,bool set_content_len = true);
-    /**
-     * 判断http客户端是否有权限访问文件的逻辑步骤
-     *
-     * 1、根据http请求头查找cookie，找到进入步骤3
-     * 2、根据http url参数(如果没有根据ip+端口号)查找cookie，如果还是未找到cookie则进入步骤5
-     * 3、cookie标记是否有权限访问文件，如果有权限，直接返回文件
-     * 4、cookie中记录的url参数是否跟本次url参数一致，如果一致直接返回客户端错误码
-     * 5、触发kBroadcastHttpAccess事件
-     * @param path 文件或目录
-     * @param is_dir path是否为目录
-     * @param callback 有权限或无权限的回调
-     */
-	void canAccessPath(const string &path,bool is_dir,const function<void(const string &errMsg,const HttpServerCookie::Ptr &cookie)> &callback);
-
-    /**
-     * 获取用户唯一识别id
-     * 有url参数返回参数，无参数返回ip+端口号
-     * @return
-     */
-	string getClientUid();
 
 	//设置socket标志
 	void setSocketFlags();
